Correção de alguns bugs visuais em algumas partes do site, menu com alguns jogos passando

+++ LocadoraDejogos/Models/Jogos.cs
@@ -47,7 +47,7 @@ namespace LocadoraDejogos.Models
         public int? Unidade { get; set; }
 
         [DisplayName("Preço")]
-        public int? Preco { get; set; }
+        public float? Preco { get; set; }
